+++ opensearch/src/client.rs
@@ -104,3 +104,13 @@ impl OpenSearch {
             .await
     }
 }
+
+#[cfg(test)]
+mod test {
+    #[test]
+    fn client_is_send_sync() {
+        // Ensure that the client is `Send` and `Sync`
+        fn is_send_sync<T: Send + Sync>() {}
+        is_send_sync::<super::OpenSearch>()
+    }
+}